Extract Tile Bus protocol + ITransport into a shared library
src/row/ was about to become a second, independently-maintained
implementation of the exact wire format src/tile/DF2-Tile already has
(Frame/Cmd/CRC/FrameParser + ITransport). Since both live in the same
repo, extract them to src/common/tile_bus_protocol/ instead and have
both projects' platformio.ini reference it via lib_extra_dirs, so
there's one canonical implementation rather than two kept in sync by
hand.

Frame/Cmd/FrameParser/crc16/frame_encode/ITransport are unchanged from
the tile project's original names - no renames needed on that side.
lib/df2_core/command_handler.h needed an explicit relative include
since PlatformIO's LDF doesn't propagate lib_extra_dirs include paths
into another lib/ folder's own compile step.

Verified: all four tile-project PlatformIO envs build, the Makefile-based
tools/tile-bus-broker builds, test/integration/test_sense_mapping.py
passes end-to-end, and all four row-project envs still build with the
added lib_extra_dirs.
