+++ src/row/platformio.ini
@@ -13,6 +13,9 @@
 ; Each named environment appends only its own entry point via ${env.build_src_filter}.
 [env]
 build_src_filter = +<*> -<**/main*.cpp>
+; tile_bus_protocol/ is shared with src/tile/DF2-Tile/ (both sides of the same
+; wire protocol) - see src/common/tile_bus_protocol/.
+lib_extra_dirs = ../common
 
 [env:seeed_xiao_rp2350]
 platform  = https://github.com/Seeed-Studio/platform-seeedboards.git
