+++ src/tile/DF2-Tile/include/transport.h
@@ -1,13 +1,4 @@
 #pragma once
-#include "protocol.h"
-
-class ITransport {
-public:
-    virtual ~ITransport() = default;
-    virtual void init() = 0;
-    // Consume one available byte and feed it to parser. Returns true (and fills
-    // *out) when a complete valid frame is received. Call in a loop until false
-    // to drain all buffered input.
-    virtual bool poll(FrameParser &parser, Frame *out) = 0;
-    virtual void send(const Frame &frame) = 0;
-};
+// Implementation lives in src/common/tile_bus_protocol/ so it is shared with
+// src/row/ (the row controller is the other side of this same wire protocol).
+#include "../../../common/tile_bus_protocol/transport.h"
