+++ src/tile/DF2-Tile/platformio.ini
@@ -13,6 +13,9 @@
 ; Adding a new main_foo.cpp test harness requires no changes to other environments.
 [env]
 build_src_filter = +<*> -<main*.cpp>
+; tile_bus_protocol/ is shared with src/row/ (both sides of the same wire
+; protocol) - see src/common/tile_bus_protocol/.
+lib_extra_dirs = ../../common
 
 [env:ATtiny3224]
 platform = atmelmegaavr
